% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewApproachRawData.R
\name{getDoc}
\alias{getDoc}
\title{Function to access RawData texts}
\usage{
getDoc(x, i)
}
\arguments{
\item{x}{RawData object}

\item{i}{index}
}
\value{
returns text
}
\description{
Function to access RawData texts
}

