% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewApproachRawData.R
\name{getMeta}
\alias{getMeta}
\title{Function to access RawData metadata}
\usage{
getMeta(x, i, meta)
}
\arguments{
\item{x}{RawData object}

\item{i}{index}

\item{meta}{metadata field name}
}
\value{
returns metadata
}
\description{
Function to access RawData metadata
}

