% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewApproachRawData.R
\name{NewRawData}
\alias{NewRawData}
\title{Function to create}
\usage{
NewRawData(x, language)
}
\arguments{
\item{x}{the source of}

\item{language}{default language of document}
}
\value{
returns RawData list
}
\description{
Function to create
}

