% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpTm.R
\docType{package}
\name{tmpTm}
\alias{tmpTm}
\alias{tmpTm-package}
\title{tmpTm: A Text mining combo}
\description{
At the beggining is just simple package, later I will change it for
some more.
}
\section{nowa}{

cos tu sie wpisuje
}

